instr: Implemented AndOperator and its simplification
Two new mechanisms were added: absorber (element that, if present,
becomes the result value, like 0 in multiplication) and idempotency
(the operation applied to two instances of the same element yields
that element).

Two test cases fail at the moment. One of them requires more invasive
changes than I want to include in this commit and for the other I'm
not sure yet what the most elegant way is to resolve it.