instr: Improved delayed error handling
Use a "with" block to explicitly control where delayed error handling
is performed. This is more readable and puts the decision of whether
to continue or not in the hands of the caller rather than the callee.

Also reduced the amount of code needed for checking duplicate
arguments. It now reports a duplicate as soon as it is detected,
which is an improvement. It also reports triplicates etc again for
each argument after the first, which is unlikely to make any
difference in practice.