instr: Implemented removal of redundant store nodes
Two test cases were added: the first (same value) already passed on
the pre-existing code, but that feature wasn't tested yet. I can't
think of a scenario in which the second (different value) test would
pass and the first wouldn't, but if the first passes and the second
does not that provides useful information about where the bug might
be, so I think it is useful to keep the first test.