instr: Fixed issues reported by pylint
Most are harmless like PEP 8 whitespace violations and too-short
variable names. Also I had to locally suppress a few checks that were
reporting false positives.

There was one actual bug: "canLoadHaveSideEffects" in LocalReference:
the final "s" is a typo.

The change in Concatenation.iterWithOffset() is a workaround for a bug
in pylint: (in the Astroid library it uses)
  https://bitbucket.org/logilab/astroid/issue/69/