instr: Do not warn about unused arguments if a function contains errors
This is to avoid false positives: errors in a function body means not
all code was emitted, which may very well be the reason why arguments
are not used. In any case, the user has something more important to
focus on.

As part of this change, Function objects will now contain None for
their 'code' attribute if there were errors in the body. I think that
will produce the best results when we deal with function calls later:
since there is a Function object, we can verify the arguments passed
to it. But a partial body might lead to false positives when trying to
merge it into the calling code.