instr: Perform constant replacement one constant at a time
The replacement map had the problem that the replacements were not
applied to constants in the replacement map itself. While it would be
possible to add code to do that, removing the replacement map is a
simpler solution. We do have to take some extra care now to not
modify a dictionary we're iterating over.

This bug is very difficult to unit test for, since we'd either have to
rely on the replacement map being processed in a particular order or
we'd have to use circular references between expressions. Both those
solutions are likely to break with future code changes. Therefore I
decided to not write a test to trigger this problem.