instr: Fixed addition/subtraction parsing
The previous code had always misparsed subtraction: "a - b - c" would
be parsed as (a - (b - c)). And since the introduction of and/or/xor,
it could also misparse addition.

In the new approach, there is no advantage anymore to having a
placeholder class for subtraction, so I removed that. The code now
converts subtractions to addition of complements when first
constructing the expression tree.