instr: Fixed constant replacement inside ComputedConstant expressions
One problem was that substitution didn't recurse into the expressions
of ComputedConstants. A custom implementation of the substitute()
method solves this.

Another problem was that when replacing equal constants, Constant
objects were kept in the constsGrouped dictionary and those were not
updated on replacements. So things could break if more than one set
of equal constants was found and the second set had expressions
containing elements from the first. This was solved by storing CIDs
instead of Constant instances.