instr: Integrate local context with CodeBlockBuilder
This makes the process of importing values from the global context
more explicit. Also it is now clearer which methods are safe for
external code to call.

The removal of VariableDeclaration means that a variable declaration
on a line by itself will trigger an unused value warning. I think this
is not an issue though since declaring variables inside function call
arguments works fine, so there is never a need to have a declaration
on a line of its own.