instr: Simplified handling of extra keywords for log methods
We can let the standard Python argument handling figure it out instead
of using custom code for it.

Copy the location instead of modifying it: there could be multiple
column spans applied to the same full-line location.

Allow creating a location dictionary that already contains the column
span information. This is not used yet, but might be useful to produce
precise error messages about identifiers that are discovered to be used
in an invalid way after the parsing stage.

Also fixed a typo in the docstring of DelayedError.