instr: Process parse tree differently depending on where it occurs
In particular, the left hand side of an assignment should yield
a storage with the right hand side should yield a value.

At this stage there is not much of an advantage to this split yet,
but I think that as more code from the expression_builder and the
codeblock_builder module is merged, the storage functions will diverge
from the value functions.

Note that this removes the ability to declare a variable inside
a function argument list, but that was never there by design anyway,
just something that emerged from the implementation.