instr: Split off declaration from definition in parse tree
The ability to declare separately from definition will be useful when
functions will be returning references: in that case "ret" should be
defined by the function body, but its type is already declared in the
function header.

The code in expression_builder that processes the separated parse nodes
is rather ugly at the moment. This will be improved in a future commit.