instr: Dropped the two-phase definition in the global context
It was originally introduced as a way to get name collisions reported
on the right input line, but passing a location object is an
alternative way to achieve the same goal.

It might be useful to split declaration from definition at some point,
but for now it is simpler to have just one call to add items to the
context.