instr: Always store sequence in inlineBlock's ridMap
There used to be special treatment for LocalReferences, but while
the code for other references was simpler, having one simple and
one complex code path is still more complex than having only one
code path.

Changed sequence from list to tuple, to make explicit that we're not
going to modify it.

In the test case, the register width was corrected: default width is
8 bits, but the test value was 16 bits wide.