instr: Have inlineBlock() take ComposedStorages for reference arguments
The only non-test caller is _convertFunctionCall(), which was in fact
already passing ComposedStorages for reference argument values ever
since buildStorage() has been returning ComposedStorages. So that was
a bug that is fixed by this commit. The test code wasn't passing
ComposedStorages yet and had to be updated.

The test_storage_decompose module is in a weird spot now, since a lot
of the code it is testing is only used by those same tests. It could
use a redesign.