instr: Remember location in Load/Store nodes
Previously, the location was remembered in a separate loadLocations
administration in the code block builder. This the new approach also
works when a load is emitted as part of a partial store.

By passing location information to methods, we get a more accurate
location including span information, instead of just the current line.