instr: Simplify (A + $10001)[:16] to (A + 1)[:16]
This pattern can occur for SP when analyzing this function:

= func test()
    var u16 X
    pop(X)
    push(X)

Here, (A - 1)[:16] becomes (A + $FFFF)[:16]) and when adding 2 later
we get (A + $10001)[:16].

Structurally, (A + $10001) and (A + 1) are considered equally complex,
since both are a variable plus a literal, but we prefer the latter.

Also try term simplification one term at a time, keeping any
improvements, rather than require improvements for the whole addition.