instr: Store ComposedStorages in global context
This required big changes in the importing, since a rid renumber is
necessary and for FixedValue the constant must be imported and its
cid renumbered as well.

I expect several cleanups are possible now, but this change is large
enough already, so let's commit it as soon as it passes the tests.