instr: More aggressive simplification in generic stages
In particular the interactions between flattening trees of associative
operations and subexpression simplification are exhausted now.

Also fixed a bug that wasn't triggered before: the custom AND
simplification code breaks if there is only one term and that term
is a literal. Since it cannot do anything useful with fewer than
2 terms anyway, it now returns immediately when called with 1 term.