instr: Make IOChannel untyped
Specify the element and address size as a width in bits, not as an
integer type.

The elements are just bit strings, which could be used as integers
but could also be used as other data types if we supported those.
For example, if there would be floating point values, those could be
read from and written to memory.

For tracing values, an array-like storage such as RAM needs to be
indexed with discrete values. We could support a lookup using a
floating point index as is used for textures, but that would just
be mapped implicitly to a discrete index when emitting the code
block: at the level of the I/O channel we'd use integers.

Since we're only ever going to use integers for both the address and
the data, wrapping the widths in IntType is inefficient at best and
misleading at worst.