instr: Make Storage untyped
See the preceding IOChannel commit for motivation.

Note that NamedStorages keep their typing, since they play a dual role:
they are both storages and context elements and we do want types on
context elements such as registers.

Storages that are not NamedStorages, being IOReference and FixedValue,
now only have a width and not a type. Also the "type" property for
ComposedStorage was dropped.