instr: Define constants as FixedValues instead of plain constants
A FixedValue is a storage and therefore it can preserve the declared
width, while a plain constant only has the expression width which can
change with simplification (and which I want to throw out).

Another advantage is that there is now less difference between defining
a constant and defining a reference.