instr: Fixed handling of declarations and definitions inside expressions
Variable declaration on the value side (rhs, index, value arg etc) is
now explicitly disallowed. Previously it was unhandled, which let to
an assert being triggered.

Variable declaration on the storage side (lhs, reference arg etc) is
useful and therefore allowed; no code related to this was changed.

Constant and reference definitions inside expressions are now
disallowed; it was already disallowed on the value side, now also on
the storage side. It would be easy to allow it, but I don't see any
point to it: it only enables hard to read code in the instruction set
definition.

Note that the previous definition handling code was broken: "node.kind"
is a non-existing property: ever since the split between the definition
and declaration nodes, "kind" is a property of the declaration.