instr: Allow unlimited-width storages
Expressions are allowed to have unlimited with, but until now storages
were not. I didn't have a real reason for not allowing it, other than
a general rule of not adding a feature unless there is a good use case
for it. Allowing unlimited-width storages allows wrapping arbitrary
values, in particular results of _convertArithmetic(), in FixedValues.