instr: Add "mask" property to Expression
The property returns a bit mask which indicates which bits in the
expression value might be set.

A unit test is included and all new tests pass. An older unit test
fails now with infinite recursion, but I'd rather not spend time on
that since the code that is causing it is on the list to be rewritten
soon. Similarly, I didn't write unit tests for truncation, since I
expect the Truncation class to disappear soon.

A more precise mask is possible for at addition and complement; that
will be implemented in separate commits.