instr: Removed specialized AND simplifier for trailing zeroes in masks
This was triggering an infinite recursion in one of the unit test
cases, while all tests pass after removing the code.

I will revisit the mask-based simplification code, since I'm not
certain it tries all feasible simplifications in its current form,
but if anything is missing it will be added in the per-subexpression
simplificaton loop instead of as an operation on all subexpressions.