instr: Separate dropping of masked expression parts from truncation
By using a full mask instead of a truncation width, irrelevant parts
can be dropped also if they're not at the front of the bit string.

This also reduces the number of places where Truncation instances are
created.

Expected results in several unit tests were updated. The previous
tests considered (L[:6] + 2)[:6] to be simpler than (L + 2)[:6] and
(-L[:6])[:6] to be simpler than (-L)[:6], which is neither intuitively
true nor backed up by the complexity() function. A third test was
expecting a correct result, but the new code returns a different but
also correct result.