instr: Removed type information from Expression
Only integer types are currently supported, so when I write "type
information", I am actually referring to the width of the type in bits.

In some cases the width referred to the declared width of a value while
in other cases it referred to the actual width determined by analysis;
this ambiguity was making it hard to read code. In the new situation,
storages have a declared width that is also their actual width and
expressions do not have a declared width.

The actual (analyzed) width of expressions is now handled via the mask,
which is more powerful than just a width since it can deal with for
example trailing zeroes as well.

The lack of type information on ConstantValues is blocking some
simplifications that could be made in code blocks. Possible solutions
would be inlining expressions or propagating masks.