instr: Removed evalExpr() for now
This function was introduced to evaluate the encoding and semantics
field of mode tables, but no longer used after I decided to only
parse and not evaluate modes for now.

The reason for removing it is that adds extra complications to the
introduction of signed types. I'd rather re-add it once the rest of
the signed types support is working.