instr: Made test method addRegister() return BoundReference
It used to return a SID, but I want all load/store emission to happen
through BoundReferences, as a preparation for signed integer support.

The test_duplicate_register test case was removed since there were no
duplicate registers added in the first place, so it wasn't actually
testing the functionality of removeDuplicateStorages().