instr: Made emitRegister() return a BoundReference
Neither the main code nor the test code uses the returned object,
but for consistency with the other emit methods it should return
a BoundReference anyway.

Note that TestCodeBlockBuilder.addRegister() can't use the returned
BoundReference, since that contains the SID in the global context,
while it will be used in the local context.