instr: Don't slice loaded/stored values unless necessary
While the redundant shift and AND operations will be removed by
a later simplification pass, this takes up a significant amount
of time, which we can save by not introducing them in the first
place.

This change makes the unit tests run about 30% faster (the total
execution time, not just tests that actually hit the optimized
code).