instr: Resolved TODO in CodeBlockBuilder.inlineBlock()
It is safe to use a temporary unsigned reference to a copied storage,
even if the original reference to it was signed. Added a comment that
explains why and unit tests to verify it.

The original code always used a signed reference, which is incorrect
and will cause one of the new unit tests to fail.