instr: Parse context using expression parser
Now we can support definitions in context elements, which is useful
for modeling for example relative addressing (A=pc+N). The value is
not used yet (see TODO), but it is parsed properly.

Because the value part of a definition is an expression, it can
contain commas between function call arguments. So splitting the
context elements on commas, like our previous simplistic parser did,
no longer works.