instr: Replace BoundReference class with Reference hierarchy
The BoundReference design was blocking support for non-literal slice
indices.

The direct problem was that bit offsets were stored as Python
integers, which is too limiting. We need to support offsets that are
decoded from the instruction opcode. We might even need to support
offsets that are determined by a state inside the described processor.

But there was also a more fundamental indirect problem. By flattening
reference slicing on parsing, it is possible that when references are
concatenated and then sliced, some of the concatenated references are
dropped completely from the final composition. This means that no
load/store nodes will be emitted for them.

However, it could be difficult to tell which subreferences will be
dropped by just looking at the instruction set definition. And if we
would add support for non-literal offsets to the BoundReference
design, whether a reference would be dropped could depend on whether
the expression simplifier can reduce an expression to a literal.
So the semantics of reference composition would be complex and
implementation-dependent, which are both undesired properties.

In the new design, the reference composition is put into a data
structure which mirrors the parse tree. When storing a value into
a slice, all subreferences are first loaded, then the relevant slice
of the value is updated and finally all subreferences are stored to.
It is up to later optimization passes to eliminate loads and stores
that have no effect.

TODO: Fix unit tests.