instr: Update decompose tests to work with new references design
Since it is now references and not storages that are decomposed,
I renamed the test.

In the new references design the decomposition doesn't happen when
constructing reference objects, but when accessing them through load
or store operations. So I changed the tests to generate load code
and then check whether the generated code matches the expected
decomposition.

The test infrastructure to perform the checks is non-trivial.
I considered testing only the created reference data structure, but
the complexity of the code under test is in the load/store emission
and no longer in the construction of the reference data structure.

Another way to reduce the amount of test infrastructure would be to
run the simplifiers on the code blocks and expressions generated,
but that would pull in a large amount of complex code under test.
I decided that keeping this test specific was more important than
keeping it small.

TODO: Test store code generation as well.