instr: Made FixedValue into a Reference instead of Storage
Having it as a Storage led to unnecessary extra work.

The only advantage it had was that it allowed the "ret" variable to
be treated as a reference regardless of whether the return type is
a value or a reference. Maybe there is a different way to accomplish
this again.