instr: Removed support for appending to an already defined mode
Originally I thought this was a simple feature that introduced a lot
of flexibility for little cost. However, it would at least require
detection of circular inclusion of modes, to avoid unending recursive
lookups. So there is a non-trivial cost.

I am also unsure whether the feature is worth having. It is already
possible to extend a mode by including it in another mode, under
a new name. By using that new name, modes and instructions make an
explicit choice to use the extended version, which avoids surprises.

There was one use of mode appending in the definition files I made
so far, but it was a case of an obsolete definition co-existing with
a refactored one. So the feature was hiding a flaw in the definition
file.