instr: Gave up on idea of having slices with undetermined width
Z80 and 6502 only shift a fixed number of bits at a time (1 or 4),
but for exampple MIPS has instructions where the number of bits to
shift is determined by an immediate or register. There are already
Expression subclasses to support that (LVShift and RVShift), but
there was no support yet in the definition language.

I thought the definition language might not need shift operators and
instead use slicing to perform all shifts. This would require the
use of slices where the width is not fixed. I tried using 'int&' as
the type for slices where the width is unknown and while right-shifts
can be expressed then, left-shifts cannot, since 'A ; 0[:N]' would
have an unlimited-width integer literal in the non-first position
of the concatenation.

So I decided to introduce explicit shift operators after all.