instr: mips: Add almost complete definition for MIPS I instruction set
Multiplication, division and modulo are missing, since those are not
in the language yet.

Branch delay slots are missing; those would either require explicit
support to be added or the instruction decoding to be defined as part
of the instruction set.

Load delay slots are missing, but we can ignore those until we start
generating code, which is a long way off.

Exceptions are not modeled yet; again this would require either
explicit support or modeling of the instruction decoding.