instr: Disallow space between type name and '&' in declarations
The language never allowed whitespace there, but the implementation
did, until now.

The root of the problem is that we want '<name>&' to mean an identifier
followed by a bitwise AND operator when used as part of an expression,
but in a declaration we want '<name>&' to mean a reference type. Around
an operator we want to allow whitespace, but between a type name and
the '&' character we do not.

The solution I picked was to check for whitespace using the column
spans of the identifier and the '&' token. This is ugly conceptually
but reliable in practice. A proper solution would require using a
separate tokenizer or a new parameter to indicate what kind of token
we expect at a certain point in the parsing, but that seems like more
trouble than it's worth.