instr: Fixed import of FixedValue into local namespace
Previously, an imported FixedValue kept using the block and CID it
was imported from. While this didn't seem to be causing problems yet,
it would have broken sooner or later. Especially since, when loaded,
FixedValue returns a ConstantValue that uses the parent CID but does
not contain a block reference, so is implicitly assumed to be in the
child block (the block the namespace belongs to).