instr: Use separate code block builder for encoding field of modes
Previously I used a single builder for both the encoding and semantics
fields, but that makes no sense if you look at it closer: placeholders
for mode entries should represent the corresponding field from the
included mode; having the semantical interpretation substituted in
the encoding expression is nonsense.

I recycled the code for the obsolete ModeCodeBlockBuilder class to
make the new EncodingCodeBlockBuilder class.