instr: Reject context definitions (not declarations) for modes
Such a definition would act as a filter on the mode. While this sounds
like a useful feature, it was only used once in the current definition
files and it was easy to alter the definition to not need it.

I decided it was better to disable the feature than to implement it
right now. It could be implemented later.