instr: Improved error handling of mode context parsing
The previous code would log and continue on some errors and abort
the context parsing on other errors. Now we log and continue on all
errors, so we can catch as many errors as possible on each run.

Since we skip parsing of the other fields when an error is found in
the context, there is no point in preventing errors reported by the
semantics builder from being added to encErrors, as encErrors will
not be used if parsing of the encoding field is skipped.