instr: Look up constructor signature only once per class
Profiling checkdef showed that calls to signature() took up more than
half of the execution time. It seems it is an expensive function and
we called it thousands of times.

We only use signature() on the constructors of Expression subclasses,
so I made a metaclass for Expression that performs the signature()
call once and stores the result.

I had to remove the Singleton metaclass from Unit, since it cannot
have a metaclass itself and an unrelated metaclass from its parent.
Since we only instantiate Unit once anyway, this doesn't make a
difference in practice.