instr: mips-i: Removed mode table entry for NOP
Defining the all-zero opcode as "nop" has the unwanted side effect
that "sll zero,zero,0" is no longer considered a valid instruction:
"nop" has the same encoding and therefore the later definition
completely replaces the earlier one.

Instead of adding alternative syntax and pseudo-instructions as
entries to the mode tables, we need explicit support for them.