instr: Use 'encoding' instead of 'opcode' in documentation
The definitions I found for 'opcode' describe it as the part of an
instruction that selects the operation to perform, so excluding the
operands. What I need is a term that describes the whole. Since the
code uses the word 'encoding', I changed the documentation to use
that as well.

Also updated the part about instruction widths: we do support other
widths than 8-bit now.