instr: Made inclusion of encoding elements after the first explicit
There was no way yet to tell where additional elements should go;
just putting them at the end of the encoding is not always what the
instruction set needs.

I considered having 'append' as the default when no explicit include
is used, but it is better to force the user to make a conscious effort
in placing included encoding elements, to avoid mistakes. After all,
I spotted a few mistakes in the Z80 definition when I checked whether
the default behavior would do what I had expected it to.

This commit only updates the documentation; the implementation will
be updated in separate commits.