instr: Inlined _checkEncodingWidth()
There was only one caller left and the interface wasn't very elegant,
so I figured it didn't have to remain a separate function.

Along the way, fixed a bug where empty encodings were not checked for
errors. This got broken in the commit that removed the u0 encoding
type for empty encoding sequences.