instr: Add support for parsing encoding multi-matches ("X@")
For now, multi-matches only match auxiliary encoding elements. Later
support should be added for matching the full sequence when there is
no placeholder for the first element.

Decoding of multi-matches is not implemented yet. See added TODO.

There are no errors and/or warnings issued yet for missing multi-
matches in entries that include modes with auxiliary encoding
elements.