instr: No longer allow absent zero-width placeholders
When empty encoding sequences had type u0, this exception was useful.
But now that empty encoding sequences are considered to not have a
type at all, there is no point to keep it: there is no practical
reason to have u0 as an encoding type, but if the user nevertheless
chooses to create a definition that has one, they can still make it
work by adding a placeholder for it.