instr: Avoid duplicate error report when "X@" is used in expression
Since the expression is handled as a single match, the MultiMatchNode
is not claimed as a multi-match, so when checking whether all aux
elements are matched it was reported again as an error.

To avoid this, when building the list of multi-matched names, we look
for MultiMatchNodes at all levels of the expression parse tree instead
of just the top level.