instr: Simplified passing of location info to log formatter
The InputLocation class was introduced after much of the rest of the
logging code was already written. As a result, InputLocation was
written to provide what that existing code needed.

This commit instead takes InputLocation as the starting point and
rewrites the other code accordingly. This means we now pass a single
key-value pair containing the location object as extra logging info,
where before we split the location data over four pairs.

A concrete advantage is that the __slots__ feature now actually
avoids having a dictionary in every InputLocation object, unlike
before when the 'extra' attribute was itself a dictionary.