instr: Replace ParseNode in ValuePlaceholder.value with Expression
We do need to store the ParseNode somewhere, since we have to
process it both in the encoding and in the semantics builder.
The new PlaceholderSpec classes serve this purpose. They are only
used as part of the instruction set parsing, not in the 'mode'
module.

Store the placeholders in the ModeEntry instead of the semBuilder
namespace, since that namespace is directly available from the
builder, while reconstructing the placeholders would take
considerable effort.

Always process the semantics of the context, even if the semantics
field is being skipped, since the mnemonic processing depends on
context semantics too.