instr: Initial implementation of instruction decoding
Currently only instruction sets without auxiliary encoding elements
are supported (MIPS in practice).

Decoding of PC-relative addressing isn't working yet, but since
I wrote the MIPS definitions for that wrong there are no errors.
What happens is that the offset (branch) or lower address part
(jump) is output in the mnemonic, instead of the absolute address
that those resolve to.