asm: Replaced precomputed iterInstructionNames() by cached property
Caching means that we don't spend time computing the set unless we
actually need it. Currently only the assembler needs it.

Also originally a list was computed and only converted to a set in
the asm_formatter module, which is wasteful as instructions can have
a lot of variants and therefore occur in the list multiple times.