instr: Replace value read from program counter by Immediate
This simplifies the code that handles mode matches. In particular,
it makes the placeholder value a self-contained expression that can
be evaluated without access to the surrounding code block.

Putting the program counter value in the mapping instead of applying
it via iterMnemonic() seems like an obvious choice, but it is far
from simple to get the right address where EncodeMatch is created.
Most instruction sets need a pre-incremented address and the
increment is not known until the full instruction has been matched,
while we create EncodeMatch objects in submodes as well.