instr: Mark operands that are assigned to the program counter
These operands should become code labels when disassembling.

Previously, placeholders with defined values had those values
substituted as soon as statements were emitted on semBuilder.
This made it very hard to trace the assigned value back to the
operand.

The solution is to introduce a separate builder for the context
and in semBuilder we emit all placeholders as arguments, whether
they have a defined value or not.

Currently the semantics code block does not include the context
definitions, but that should be relatively straightforward to solve
and a TODO was inserted for it.