instr: Use lookup table for more decoding situations
If not all masks are equal, but they do share common segments, we can
create a lookup table for each such segment.

This commit also resolves a TODO about the handling of decoders that
are mapped to the same table index: those no longer prevent the use
of a table.