instr: When all decoders are assigned to the same entry, skip table
We only try to create a table if there is more than one decoder to
choose between, but it is possible that all decoders have the same
value for the selected segment. This would result in a table in
which only one entry can produce a match, in which case it is more
efficient to create a FixedPatternDecoder instead.