instr: Add roles on submode match results
This marks the operands of the LR35902's JP and CALL instructions as
code addresses. The fact that the 'NN' placeholder contained a code
address was recognized, but since 'NN' is an 'imm16' instead of 'u16'
it is a match placeholder instead of a value placeholder and the
role of the match placeholder wasn't applied to the submode match.