instr: Report error if placeholder order blocks decoding
When constructing a decoder a ValueError will be raised in this case,
but since this is a problem with the instruction set definition we
should report it during parsing. Doing so adds a lot of detail to the
error message and makes sure this issue is checked even if no decoder
is constructed.