instr: z80: Fixed and completed $DD $CB and $FD $CB instructions
The previous definitions triggered the placeholder order error that
was recently implemented. I think in theory a working decoder could
be constructed from them, but that requires separate decoders for
instructions after the $DD and $FD prefixes and I'd like to avoid
adding yet another clause to the already complex matching rules.

Also the previous definitions only defined the undocumented variants
of the RLC instruction, while all rotate and shift instructions have
these variants.

The undocumented variants that write a copy of the rotated/shifted
value to a register now do that by returning that value from the
helper function. This allows the same helper function to be used
for all variants of an instruction.

To avoid near-duplicate entries for IX and IY, I introduced the
"memx" mode for memory access using the index registers.