instr: Fix unit tests that create registers
Since registers are created in the global scope, if separate builders
add the same register, those builders should be using the same global
namespace, so they share the same register object.

I also made the type check on shared registers more strict: the full
type is compared now, instead of just the width. This way differences
in signedness are considered an error.