instr: Made Node classes mutable
Generally immutable classes lead to simpler code, but not for
simplifying code blocks: when CIDs will be removed, replacing
a Load node would require replacing the values loaded from it
in all expressions, which is a lot of work in both code and
runtime overhead. By having a mutable node, we can just update
the node and every loaded value will immediately be using the
updated version.