instr: Improve tests after constants removal
The test_return_value_renumber test was removed since what it tested
for doesn't exist anymore and it doesn't add much value besides its
core goal.

The test_return_value test was cleaned up. It is still a bit clunky
but I think it is useful enough to keep.

The test_repeated_increase test now has two variants: one which uses
a register for the counter, like the original test did, and one which
uses a local variable. The difference is that a local variable can be
optimized out completely.