instr: Avoid creating new reference objects if clone is identical
When importing SingleReferences, the code block must be replaced, but
tests show there are sufficient cases where cloned references are
identical to the original and we can avoid creating new objects in
that case.

Having this functionality in Reference.clone() also obsoletes the
dedicated change check mechanism in CodeBlock.updateRetRef().