instr: Treat placeholder value more like a function argument
Functions have a local variable for each value argument, which is
initialized to the passed value, but can be changed by the function
body.

Now the semantics field works in a similar way. Even though there
can be only one statement in the semantics field, there is an
observable difference when a reference to a placeholder is passed
to a function: previously anything written to that reference would
be discarded, while now it persists (with a life time limited to
a single semantics evaluation).

Since the encoding builder is not allowed to emit any loads, that
will still emit a FixedValue instead of a local variable.