instr: Fix computation of context-defined placeholder value
This was broken since dd9e7aeb (Treat placeholder value more like
a function argument). That commit puts a local variable into the
namespace rather than the argument itself, which resulted in
EncodeMatch not being able to simplify the placeholder value into
an IntLiteral.

Most of the diff is in moving the placeholder construction code
to combine it with the emitting. The actual change is the use of
createCodeBlock() with an overridden return reference name.