disasm: Split off mnemonic generation into new class
The parts needed for decoding remain in EncodeMatch, while the parts
needed for generating the mnemonic representation of a match are now
in the new InstructionMatch class.

This gets rid of the setPC() method: since InstructionMatch is
instantiated by the disassembler, it can pass the value of the
program counter as a constructor argument instead.

It also reduces the amount of code in the 'mode' module, which has
grown quite large and complex.

For the future, I expect more semantic analysis to be added in
InstructionMatch. This analysis is not necessary for other uses of
the instruction decoder, for example an assembler checking for
overrides. So this split avoids the overhead of unnecessary analysis
without having to take special care to perform it on-demand.